refactor: extract sop template api
